requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 24.11.3
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=0.9.0=pypi_0
adjusttext=1.3.0=pypi_0
anndata=0.8.0=pypi_0
anyio=4.5.2=pypi_0
argon2-cffi=25.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
astor=0.8.1=pypi_0
asttokens=3.0.0=pypi_0
astunparse=1.6.3=pypi_0
async-lru=2.0.4=pypi_0
attrs=25.3.0=pypi_0
babel=2.17.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.13.4=pypi_0
biothings-client=0.4.1=pypi_0
bleach=6.1.0=pypi_0
blosc2=2.0.0=pypi_0
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.6.15=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.2=pypi_0
comm=0.2.2=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=10.0.130=0
cudnn=7.6.5=cuda10.0_0
cycler=0.12.1=pypi_0
cython=3.1.2=pypi_0
debugpy=1.8.14=pypi_0
decorator=5.2.1=pypi_0
defusedxml=0.7.1=pypi_0
exceptiongroup=1.3.0=pypi_0
executing=2.2.0=pypi_0
fastjsonschema=2.21.1=pypi_0
fonttools=4.57.0=pypi_0
fqdn=1.5.1=pypi_0
gast=0.3.3=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.24.3=pypi_0
h11=0.16.0=pypi_0
h5py=3.11.0=pypi_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.5.0=pypi_0
importlib-resources=6.4.5=pypi_0
ipykernel=6.29.5=pypi_0
ipython=8.12.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.2=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
json5=0.12.0=pypi_0
jsonpointer=3.0.0=pypi_0
jsonschema=4.23.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter-client=8.6.3=pypi_0
jupyter-core=5.8.1=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.2=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.3.8=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.3=pypi_0
keras=2.2.4=pypi_0
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
legacy-api-wrap=1.4.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libxcb=1.17.0=h9b100fa_0
llvmlite=0.41.1=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.7.5=pypi_0
matplotlib-inline=0.1.7=pypi_0
mistune=3.1.3=pypi_0
msgpack=1.1.1=pypi_0
mygene=3.2.2=pypi_0
natsort=8.4.0=pypi_0
nbclient=0.10.1=pypi_0
nbconvert=7.16.6=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
networkx=3.1=pypi_0
notebook=7.3.3=pypi_0
notebook-shim=0.2.4=pypi_0
numba=0.58.1=pypi_0
numexpr=2.8.6=pypi_0
numpy=1.22.4=pypi_0
nvidia-pyindex=1.0.9=pypi_0
nvidia-tensorflow=1.15.5+nv23.03=pypi_0
openssl=3.0.16=h5eee18b_0
opt-einsum=3.3.0=pypi_0
overrides=7.7.0=pypi_0
packaging=25.0=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
patsy=1.0.1=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.4.0=pypi_0
pip=24.2=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.3.6=pypi_0
pot=0.8.2=pypi_0
prometheus-client=0.21.1=pypi_0
prompt-toolkit=3.0.51=pypi_0
protobuf=3.20.3=pypi_0
psutil=7.0.0=pypi_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.3=pypi_0
py-cpuinfo=9.0.0=pypi_0
pycparser=2.22=pypi_0
pygments=2.19.2=pypi_0
pynndescent=0.5.13=pypi_0
pyparsing=3.1.4=pypi_0
python=3.8.20=he870216_0
python-dateutil=2.9.0.post0=pypi_0
python-json-logger=3.3.0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
pyzmq=27.0.0=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.32.4=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rpds-py=0.20.1=pypi_0
scanpy=1.9.3=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.11.2=pypi_0
send2trash=1.8.3=pypi_0
session-info=1.0.1=pypi_0
setuptools=75.1.0=py38h06a4308_0
setuptools-scm=8.3.1=pypi_0
sinfo=0.3.4=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.7=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.1=pypi_0
stdlib-list=0.10.0=pypi_0
tables=3.8.0=pypi_0
tensorboard=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
tinycss2=1.2.1=pypi_0
tk=8.6.14=h993c535_1
tomli=2.2.1=pypi_0
tornado=6.4.2=pypi_0
tqdm=4.67.1=pypi_0
traitlets=5.14.3=pypi_0
trvae=1.1.2=pypi_0
types-python-dateutil=2.9.0.20241206=pypi_0
typing-extensions=4.13.2=pypi_0
tzdata=2025.2=pypi_0
umap-learn=0.5.7=pypi_0
uri-template=1.3.0=pypi_0
urllib3=2.2.3=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=24.8.0=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.6=pypi_0
wheel=0.44.0=py38h06a4308_0
wrapt=1.11.1=pypi_0
xlrd=1.2.0=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.20.2=pypi_0
zlib=1.2.13=h5eee18b_1
